+++ lib/src/catch_response.dart
@@ -3,5 +3,4 @@ import 'package:http_interop_middleware/http_interop_middleware.dart';
 
 /// Middleware that allows the inner handler to throw a Response.
 final Middleware catchResponse =
-    middleware(onError: (err, trace) async => (err is Response) ? err : null);
-
+    middleware(onError: (err, t, r) async => (err is Response) ? err : null);