+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+## [0.1.0] - 2024-07-14
+### Added
+- `Middleware.add()` extension method
+
+### Changed
+- Added the Request to the callbacks in `middleware()`
+
 ## [0.0.1] - 2024-07-13
 ### Added
 - initial
 
+[0.1.0]: https://github.com/f3ath/dart-http-interop-middleware/compare/0.0.1...0.1.0
 [0.0.1]: https://github.com/f3ath/dart-http-interop-middleware/releases/tag/0.0.1